pyproject.toml
[project]
name = "poc-multiple-entity-openai-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.69",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
]
